bots/weather_bot/requirements.txt
python-dotenv==1.0.1
python-telegram-bot==13.15
requests==2.32.2
